+++ lib/state.sh
@@ -184,6 +184,8 @@ apo_state_initialize() {
     apo_state_set RECOVERY_WAIT_CONTEXT ''
     apo_state_set RECOVERY_WAIT_STARTED_AT ''
     apo_state_set RECOVERY_WAIT_TIMEOUTS 0
+    apo_state_set TRANSIENT_RETRY_CONTEXT ''
+    apo_state_set TRANSIENT_RETRY_COUNT 0
     apo_state_set MANUAL_TEST "${APO_MANUAL_TEST:-0}"
     apo_state_set MANUAL_CPU "${APO_MANUAL_CPU:-}"
     apo_state_set MANUAL_GPU "${APO_MANUAL_GPU:-}"